fix: common functions to update, remove and filter elements form list
