+++ src/pages/Appearance/Programs/ProgramSpecificSettings.js
@@ -6,8 +6,9 @@ import isEqual from 'lodash/isEqual'
 import PageSubtitle from '../../../components/page/PageSubtitle'
 import NewProgramSpecific from './NewProgramSpecific'
 import SpecificTableAction from './SpecificTableAction'
-import { filterUnusedElements, prepareSpecificSettingsList } from './helper'
+import { prepareSpecificSettingsList } from './helper'
 import { useReadProgram } from './programQuery'
+import { filterUnusedElements } from '../../../utils/utils'
 
 const ProgramSpecificSettings = ({
     onChange,
