+++ src/pages/Synchronization/Programs/NewProgramSpecific.js
@@ -13,10 +13,10 @@ import {
     createInitialSpecificValues,
     findProgramNameById,
     isProgramWithRegistration,
-    updateSettingsList,
 } from './helper'
 import { populateProgramObject } from '../../../modules/programs/populateProgramObject'
 import { parseValueByType } from '../../../modules/programs/parseValueBySettingType'
+import { updateSettingsList } from '../../../utils/utils'
 
 const NewProgramSpecific = ({ programList, rows, handleRows, disabled }) => {
     const [openDialog, setOpenDialog] = useState(false)
