+++ src/utils/utils/index.js
@@ -1,0 +1,3 @@
+export * from './filterUnusedElements'
+export * from './removeSettingsFromList'
+export * from './updateSettingsList'
