+++ src/utils/utils/removeSettingsFromList.js
@@ -1,0 +1,6 @@
+export const removeSettingsFromList = (setting, settingList) =>
+    settingList.filter(element => {
+        const elementID = element.uid || element.id
+        const settingID = setting.uid || setting.id
+        return elementID !== settingID
+    })
